Added getting orders info functionality for admins

+++ SunnyBookShop/Services/AdminService.cs
@@ -3,6 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using SunnyBookShop.Interfaces;
 using SunnyBookShop.Models;
 using SunnyBookShop.Persistence;
+using SunnyBookShop.ViewModels;
 
 namespace SunnyBookShop.Services;
 
@@ -11,10 +12,10 @@ public interface IAdminService
     Task<Book> AddBookAsync(Book book);
     Task<Book?> EditBookAsync(int id, Book book);
     Task<bool> DeleteBookAsync(int id);
+    Task<OrdersViewModel> GetOrdersAsync();
 }
 
-
-public class AdminService: IAdminService
+public class AdminService : IAdminService
 {
     private readonly IPhotoService _photoService;
     private readonly ApplicationDbContext _dbContext;
@@ -30,15 +31,16 @@ public class AdminService: IAdminService
     public async Task<Book> AddBookAsync(Book book)
     {
         UploadResult uploadResult;
-        
-        if(book.PosterFile is not null)
+
+        if (book.PosterFile is not null)
         {
             uploadResult = await _photoService.UploadPhotoAsync(book.PosterFile);
             if (uploadResult.Error is not null)
                 return null;
-            
+
             book.PosterUrl = uploadResult.Url.ToString();
         }
+
         await _dbContext.Books.AddAsync(book);
         await _dbContext.SaveChangesAsync();
 
@@ -53,26 +55,40 @@ public class AdminService: IAdminService
             _logger.LogInformation($"Deletion of {book.PosterUrl}");
             var uploadResult = await _photoService.UploadPhotoAsync(book.PosterFile);
             book.PosterUrl = uploadResult.Url.ToString();
-        } 
+        }
+
         _dbContext.Books.Update(book);
         await _dbContext.SaveChangesAsync();
-        
+
         return book;
     }
 
     public async Task<bool> DeleteBookAsync(int id)
     {
         var book = await _dbContext.Books.FirstOrDefaultAsync(b => b.Id == id);
-        if(book is null)
+        if (book is null)
             return false;
-        if (book.PosterUrl is not null)
-        {
-            await _photoService.DeletePhotoAsync(book.PosterUrl);
-        }
-        
+        if (book.PosterUrl is not null) await _photoService.DeletePhotoAsync(book.PosterUrl);
+
         _dbContext.Books.Remove(book);
         await _dbContext.SaveChangesAsync();
 
         return true;
+    }
+
+    public async Task<OrdersViewModel> GetOrdersAsync()
+    {
+        var orders = _dbContext.Orders
+            .Include(o => o.Book).GroupBy(o => o.OrderGroupId)
+            .OrderBy(o => o.First().Status)
+            .OrderByDescending(o => o.First().CreatedAt);
+        var groupedOrders = await orders
+            .Select(g => g.ToList())
+            .ToArrayAsync();
+
+        return new OrdersViewModel
+        {
+            Orders = groupedOrders
+        };
     }
 }
